+++ packages/media-parser/src/index.ts
@@ -1,1 +1,9 @@
 export {parseMedia} from './parse-media';
+export {
+	AudioSample,
+	OnAudioSample,
+	OnAudioTrack,
+	OnVideoSample,
+	OnVideoTrack,
+	VideoSample,
+} from './webcodec-sample-types';
